Add possibility to start a server on wok start.
In the configuration you can specify a list or server (or supervisor)
via the key "starts". This key take a list or tuples where the first
element is an atom giving the name of the server or supervisor. The
second element is the list of parameters passed to the start function of
the server (or supervisor). The last element is the type : worker
(default) or supervisor.